Fix paged-KV decode dtype mismatch + PLE gather int64 overflow
attention_decode: PyTorch reference mixed fp16 weights with fp32 V in
matmul (weights @ v_h.float()). Fixed: keep weights as float32 from
softmax, cast output to half. GPU-VALIDATED: correct=True, max_err=
1.5e-05, 43.49 GB/s on gemma4_31b_decode_global A100.

ple_gather: tok (int32, up to 262143) * ple_layer_stride (8960) =
2.35 billion > INT32_MAX (2.15B) causing pointer arithmetic overflow
and garbage reads. Fixed: cast tok to tl.int64 and ple_layer_stride
to tl.int64 before multiplication. UNVALIDATED — Modal billing limit
hit before re-test. The int64 overflow diagnosis is high-confidence
(max_err=5.42 on random PLE values scaled 0.01x is consistent with
reading from a wrapped-around address).

556/556 local tests still passing.

+++ src/research_engine/triton_attention_decode.py
@@ -393,10 +393,10 @@ def torch_paged_attention_decode(q, k_cache, v_cache, page_table, context_lens,
                 v_h = v_h[start:]
 
             scores = (q_h.float() @ k_h.float().T) * sm_scale  # [ctx_len] or [window]
-            weights = torch.softmax(scores, dim=-1).half()
-            out[b, h, 0, :] = weights @ v_h.float()
+            weights = torch.softmax(scores, dim=-1)  # keep float32
+            out[b, h, 0, :] = (weights @ v_h.float()).half()
 
-    return out.half()
+    return out
 
 
 def benchmark_one(batch, num_heads, num_kv_heads, head_dim, context_len,
